Move custom tasks to separate package
