+++ src/main/groovy/org/gradle/samples/internal/tasks/SampleZipTask.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.samples.internal;
+package org.gradle.samples.internal.tasks;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.file.DirectoryProperty;
